Support c6x linking with GNU linker.
This patch adds a config option (CONFIG_TI_C6X_LINKER) which is used in the
sources to support the choice of linker used to build the kernel. If this
option is defined, the TI linker is being used to link the kernel. If this
option is not defined, the GNU linker is used.

Signed-off-by: Mark Salter <msalter@redhat.com>