add support for NAND driver and JFFS2 on EVM6457 and EVM6472
This patch adds a preliminary NAND driver for the EVM6457 and EVM6472. This
driver is a forward port of a driver from the 2.6.13 kernel. The driver itself
is a bit-banger NAND driver using GPIO. In newer kernels, there is a generic
GPIO-based bit-banger NAND driver. However, the generic driver assumes that the
NAND data port is memory-mapped on an IO bus. This is not the case for the EVM
hardware which uses GPIO also for the data lines. This driver is a first step
to just get something working. In the near future, it will be replaced with the
generic GPIO based driver once that driver has been extended to support GPIO
based data access.

Signed-off-by: Mark Salter <msalter@redhat.com>