temporary workaround for TI toolchain bug
The TI compiler (or maybe assembler) has a problem with a loop in the amdstd
FLASH command set handling. The code generated for this loop by the TI tools
gets something wrong which leads to a kernel crash. Adding a printk seems to
work around the problem for some reason. Just do this temporarily until the
true problem is isolated and resolved. This same code runs fine without the
added printk when compiled with GCC.

Signed-off-by: Mark Salter <msalter@redhat.com>