rewrite evm6488 nand driver to use generic gpio
This patch replaces the existing evm6488 NAND driver with a driver based on the
NAND gpio driver. The existing GPIO-based NAND driver was not usable because it
assumes that NAND data is accessed through a memory mapped hardware interface.
The hardware on the evm6488-like NAND interfaces use GPIO lines to drive the
NAND read-enable, write-enable, and data lines.

Signed-off-by: Mark Salter <msalter@redhat.com>