add c6x watchdog support
This patch adds watchdog support to the c6x port. The Davinci watchdog driver
is used without modification. The c6455 SoC requires that external circuitry
handle the watchdog timer output line as needed. The DSK6455 board does not
have the watchdog output connected externally. The other supported SOCs can be
configured to perform a local megamodule reset or other action when the watchdog
expires. To support this, a commandline option was added:

   wdt_action=<local|nmi|delay,NNN|system>
   where:
       local     => perform megamodule reset
       nmi       => perform an NMI (c6472 only)
       delay,NNN => nmi followed by delayed local reset (c6472 only)
                    NNN is number of CPU/6 clocks to delay. Powers of
		    two from 256 - 32768 are supported.
       system    => system reset (c6472 only)

Signed-off-by: Mark Salter <msalter@redhat.com>