C6x: Use by default the ZONE_DMA for physical memory
Use the ZONE_DMA instead of the ZONE_NORMAL for the physical memory
in order to avoid failure when allocating pages with the GFP_DMA flag.
This maybe usefull for drivers that are allocating memory using this flag
(like disk block devices). This is transparent regarding other types of
allocations (without the GFP_DMA flag).

Signed-off-by: Aurelien Jacquiot <a-jacquiot@ti.com>