C66x: Fixed PSC setup for software reboot (2)
This commit fixes the previous fix that was not fully correct.
Apparently the SGMII clock domain (if enabled) cannot be switched off correctly
if the PA clock domain is off. This causes the SGMII module to be in
an unrecoverable state.
As a consequence, only reset the PA and the SGMII clock domains if
the PA clock domain is already on.

Signed-off-by: Aurelien Jacquiot <a-jacquiot@ti.com>