KeyStone PCIe: fix PCI legacy interrupt support for using all INTx
Add the support of all the four INTx PCI legacy interrupts. Previously only
the INTA was managed and used. The management of the four INTx is needed for
bridge management (legacy interrupts are swizzled when crossing bridges).

Mapping of legacy interrupts is now shifted with the pin number and all
legacy interrupts are unmasked.

Signed-off-by: Aurelien Jacquiot <a-jacquiot@ti.com>