KeyStone PCIe: add detection of non-existent devices for C66x architecture.
Apparently the C66x does not manage the error response
to non-posted PCIe transactions with an explicit abort external
exception like on ARM with its MMU.

So in order to check if a device is present or not when
accessing the PCIe remote configuration space, we try to
read and modify the parity enable bit of the PCI command field.
According to the PCI specification this bit is r/w and can be
modified for all EP and bridges.

If the bit can be modified we come to the conclusion that
device is present. Otherwise device is not present.

This has been tested with several EP and bridges/switches.

Signed-off-by: Aurelien Jacquiot <a-jacquiot@ti.com>