KeyStone PCIe: remove call to pci_assign_unassigned_resources() in PCI driver
This call to pci_assign_unassigned_resources() must be handled by architecture
PCI part (e.g. bios32.c) because it depends if architecture has firmware/BIOS that do
this resource assignment or not. Otherwise we may call it twice or override
firmware/BIOS settings .

It may cause a problem with current ARM bios32.c that does not explicitely call
pci_assign_unassigned_resources() during PCI setup. But it is the regular location
where this function must be called (and not in the PCI driver itself).
The ARM bios32.c should be fixed in that way and calls
pci_assign_unassigned_resources() when use_firmware variable is not defined.

Signed-off-by: Aurelien Jacquiot <a-jacquiot@ti.com>