NOMMU: Fixed shared mapping in mmap() for special devices
When performing a mmap() on a already mapped (by another process) special device
the address returned by mmap() is the file offset instead of being the vm_start
resulting from the device specific mmap() method. This is because the mapping
address returned in vma->vm_start is not copied back into the result value.

This problem is apparently still present in recent kernel (3.6) and should
probably be pushed upstream.

Signed-off-by: Aurelien Jacquiot <a-jacquiot@ti.com>